Дополнили метод рендера
